# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Employee registry",
    "description": "Сервис, реализующий реестр сотрудников.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/employees/": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Получить список сотрудников",
        "description": "Возвращает список сотрудников с возможностью фильтрации и пагинации.",
        "operationId": "get_employees_api_v1_employees__get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Поиск по ФИО или телефону",
              "title": "Search"
            },
            "description": "Поиск по ФИО или телефону"
          },
          {
            "name": "is_male",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по полу",
              "title": "Is Male"
            },
            "description": "Фильтр по полу"
          },
          {
            "name": "age_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 120,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Возраст от",
              "title": "Age From"
            },
            "description": "Возраст от"
          },
          {
            "name": "age_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 120,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Возраст до",
              "title": "Age To"
            },
            "description": "Возраст до"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Номер страницы",
              "default": 1,
              "title": "Page"
            },
            "description": "Номер страницы"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Размер страницы",
              "default": 20,
              "title": "Size"
            },
            "description": "Размер страницы"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Employees"
        ],
        "summary": "Создать нового сотрудника",
        "description": "Создание нового сотрудника",
        "operationId": "create_employee_api_v1_employees__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации или сотрудник уже существует"
          },
          "422": {
            "description": "Ошибка валидации данных"
          }
        }
      }
    },
    "/api/v1/employees/{employee_id}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Получить сотрудника по ID",
        "description": "Получение информации о сотруднике по ID",
        "operationId": "get_employee_api_v1_employees__employee_id__get",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Employee Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Сотрудник не найден"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Employees"
        ],
        "summary": "Обновить сотрудника",
        "description": "Обновление данных сотрудника",
        "operationId": "update_employee_api_v1_employees__employee_id__put",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Employee Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Сотрудник не найден"
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Employees"
        ],
        "summary": "Удалить сотрудника",
        "description": "Удаление сотрудника по ID",
        "operationId": "delete_employee_api_v1_employees__employee_id__delete",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Employee Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Успешное удаление"
          },
          "404": {
            "description": "Сотрудник не найден"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EmployeeCreate": {
        "properties": {
          "last_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name",
            "description": "Фамилия"
          },
          "first_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name",
            "description": "Имя"
          },
          "patronymic": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Patronymic",
            "description": "Отчество"
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "title": "Birth Date",
            "description": "Дата рождения"
          },
          "is_male": {
            "type": "boolean",
            "title": "Is Male",
            "description": "Пол"
          },
          "phone": {
            "type": "string",
            "title": "Phone",
            "description": "Номер телефона"
          },
          "photo_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Base64",
            "description": "Фото в формате base64"
          }
        },
        "type": "object",
        "required": [
          "last_name",
          "first_name",
          "birth_date",
          "is_male",
          "phone"
        ],
        "title": "EmployeeCreate",
        "description": "Создание сотрудника"
      },
      "EmployeeListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Общее количество"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EmployeeResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "Список сотрудников"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Текущая страница",
            "default": 1
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "description": "Размер страницы",
            "default": 20
          },
          "pages": {
            "type": "integer",
            "title": "Pages",
            "description": "Всего страниц"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items",
          "pages"
        ],
        "title": "EmployeeListResponse",
        "description": "Ответ со списком сотрудников"
      },
      "EmployeeResponse": {
        "properties": {
          "last_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name",
            "description": "Фамилия"
          },
          "first_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name",
            "description": "Имя"
          },
          "patronymic": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Patronymic",
            "description": "Отчество"
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "title": "Birth Date",
            "description": "Дата рождения"
          },
          "is_male": {
            "type": "boolean",
            "title": "Is Male",
            "description": "Пол"
          },
          "phone": {
            "type": "string",
            "title": "Phone",
            "description": "Номер телефона"
          },
          "photo_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Base64",
            "description": "Фото в формате base64"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age",
            "description": "Возраст в годах"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "last_name",
          "first_name",
          "birth_date",
          "is_male",
          "phone",
          "id",
          "created_at"
        ],
        "title": "EmployeeResponse",
        "description": "Ответ с сотрудником"
      },
      "EmployeeUpdate": {
        "properties": {
          "last_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "patronymic": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Patronymic"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "is_male": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Male"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "photo_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Base64"
          }
        },
        "type": "object",
        "title": "EmployeeUpdate",
        "description": "Обновление сотрудника"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "Employees",
      "description": "Manage employees."
    }
  ]
}